Refactored common code out of pubsub/google.ts and into pubsub/google-common.ts
