+++ typescript/tests/pubsub/pubsub.test.ts
@@ -1,10 +1,5 @@
 import { parseStoreAndSend } from "../../src/pubsub/pubsub";
 import {
-    parsePayload as parseGooglePayload,
-    toDynamoEvent as googlePayloadToDynamo,
-    toSqsSubReference as toGoogleSqsEvent
-} from "../../src/pubsub/google";
-import {
     toDynamoEvent as applePayloadToDynamo,
     toSqsSubReference as toAppleSqsEvent
 } from "../../src/pubsub/apple";
@@ -12,6 +7,11 @@ import {
     StatusUpdateNotification,
     parsePayload as parseApplePayload
 } from "../../src/pubsub/apple-common";
+import {
+    parsePayload as parseGooglePayload,
+    toDynamoEvent as googlePayloadToDynamo,
+    toSqsSubReference as toGoogleSqsEvent
+} from "../../src/pubsub/google-common";
 import {HTTPResponses} from "../../src/models/apiGatewayHttp";
 import {SubscriptionEvent} from "../../src/models/subscriptionEvent";
 import Mock = jest.Mock;
